Make profile auth test work
The challenge here was that we need an AWS provider
for non-test JWT verification, but for tests, we
cannot rely on an external provider. In that case,
we should rely on a local provider that's used by
using DI.
We needed a better visibility on the errors, that's
why installing the `CallLogging` and `StatusPages`
plugins was crucial.
Cleanups will follow.